Update IAmNTimesAsOldAsYou.html
Committing a change from GPT-4 from the prompt:

You now have these two functions that more or less do the same thing, but differently:

function calculateFutureNValues(olderAge, youngerAge, olderDate, multiplier) {
... code here ...
    return futureNValues;
}

and

function calculateFutureNValues(olderAge, youngerAge, olderDate, multiplier) {
... code here ...
    return futureNValues;
}

Not sure if they ought to be combined, but the logic seems like it should be similar?  If we DID combine them and simply inserted the current date as a "divider", could that be done in a subsequent step, and could we then also simplify other functions?  Let’s work this out in a step by step way to be sure we have the right answer. 


At this point GPT-4 generated a reasonable looking generic function for both, but then started losing track of the existing state of the complete code.  So restarting with a new session after this.